backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Koresh06 <korets-24@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
environs = "^11.0.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
asyncpg = "^0.30.0"
betterlogging = "^1.0.0"
ruff = "^0.7.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^14.0.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
